baptize the newcomers

+++ test/c/test_split_2q_unitaries.c
@@ -41,7 +41,7 @@ int test_split_2q_unitaries_no_unitaries(void) {
         result = EqualityError;
         goto ops_cleanup;
     }
-    for (int i = 0; i < counts.len; i++) {
+    for (size_t i = 0; i < counts.len; i++) {
         int gate = strcmp(counts.data[i].name, "cx");
         if (gate != 0) {
             printf("gates changed when there should be no circuit changes\n");
@@ -81,7 +81,7 @@ int test_split_2q_unitaries_x_y_unitary(void) {
         result = EqualityError;
         goto ops_cleanup;
     }
-    for (int i = 0; i < counts.len; i++) {
+    for (size_t i = 0; i < counts.len; i++) {
         int gate = strcmp(counts.data[i].name, "unitary");
         if (gate != 0) {
             printf("Gates outside expected set in output circuit\n");
@@ -148,7 +148,7 @@ int test_split_2q_unitaries_swap_x_y_unitary(void) {
         result = EqualityError;
         goto ops_cleanup;
     }
-    for (int i = 0; i < counts.len; i++) {
+    for (size_t i = 0; i < counts.len; i++) {
         int gate = strcmp(counts.data[i].name, "unitary");
         if (gate != 0) {
             printf("Gates outside expected set in output circuit\n");
